#237 [fix] TodoLabelRepository 인터페이스 분리
