+++ app/src/main/java/com/gojol/notto/di/TodoLabelDatabaseModule.kt
@@ -6,6 +6,8 @@ import com.gojol.notto.model.database.TodoLabelDatabase
 import com.gojol.notto.model.database.todolabel.TodoLabelDao
 import com.gojol.notto.model.datasource.todo.TodoLabelLocalDataSource
 import com.gojol.notto.model.datasource.todo.TodoLabelLocalDataSourceImpl
+import com.gojol.notto.model.datasource.todo.TodoLabelRepository
+import com.gojol.notto.model.datasource.todo.TodoLabelRepositoryImpl
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
@@ -35,5 +37,11 @@ class TodoLabelDatabaseModule {
     @Provides
     fun provideTodoLabelLocalDataSource(todoLabelDao: TodoLabelDao): TodoLabelLocalDataSource {
         return TodoLabelLocalDataSourceImpl(todoLabelDao)
+    }
+
+    @Singleton
+    @Provides
+    fun provideTodoLabelRepository(todoLabelLocalDataSource: TodoLabelLocalDataSource): TodoLabelRepository {
+        return TodoLabelRepositoryImpl(todoLabelLocalDataSource)
     }
 }
